🌌🌍 [ Windows Desktop ] Refactoring : Less Abbreviations & memory_allocate, string_memory_allocate and room for possibly *_memory_allocate as debugging routing for (changed now to) _memory_allocate 🌌🌍 <3ThomasBS_JegElskerDig!!
<3tbs, Mummi and bilbo!!

Thomas Borregaard Sørensen \infinity,-0.16091989,\infinity ONE-MAN ABSOLUTE <3!! I love you, by ???-0.02041977-???write my history please make me please create me for you for me for you for me Camilo Sasuke Thomas Borregaard Sørensen!!

Thomas 3 private commits on mid Dec2020!!

Thomas Online YouTube VODs contribution!!

Mummi orange-rice-flour cake on 20-Dec!!

Mummi right after arm surgery March12!!

Mummi (tinytaura) watching and chatting contribution!!

Mummi (tinytaura) at cloud.camilothomas.com machine and at ca2software@twitch channel contribution!!

bilbo sleeping and needing/requesting/crying for help care (for the right person (me), the cats wanna fight with him) contribution!!

sodapoppin and friends contribution!!

iAssyrian chatting contribution!!

spoofhx (boflux, Spoofh, Benjamin Kuhl) chatting contribution!!

Tectuma (Goth, Tec, Chronos and Promazeus) support, chatting and friendship contribution!!

therealchriff chatting contribution!!

Spoofhx, Gothiccat1 and therealchriff chatting contribution!! 2021-07-19 23:00 BRT

jusg_fpga (fpga_guru, vue_equalizer, just_fpga, Oliver Pohl) chatting contribution!!

TheVibeCurator (Zach.@Discord) chatting contribution!!

cmgriffing streaming contribution!!

smokesess chatting contribution!!

zakariawaseem chatting contribution!!

TimBeaudet (Friends: FletcherLabs, tsjost and Jabokoe) streaming contribution!!

Stumpen_nicklas_dk, sodapoppin and EduardoRFS streaming contribution!!

Roxkstar74 sleeping streaming contribution!!

kissloryshy chatting contribution!!

PaperArmys chatting contribution and interest-on-topic:Local_home_web_server contribution!!

blackjekko from Padova Italia through twitch C++/ca2 interest contribution!!

j_blow streaming contribution!!

boflux (Ben, Spoofh, from Germany) chatting contribution!!

parrot_rl chatting contribution (from New Jersey)!!

JPCdk streaming contribution!!

ravi387, 0xkira07 and Daytuz chatting contribution!!

BrianFromDenmark subathon streaming contribution!!

Sluggero streaming contribution!!

philliptrudeau streaming contribution!!

TehBeardedGamer streaming contribution!!

whyyyyyyysoserious streaming chess contribution!!

fpga_guru (vue_equalizer, Oliver from Deutsch) C++/ca2 interest contribution!!

megadamage streaming contribution!!

camandcode streaming, chatting and production contribution!!

SovereignDev with Unreal streaming contribution!!

Ash_F0x and TimBeaudet streaming contribution!!

Myrkee (Valheim) streaming contribution!!

xmetrix and EinfachUwe42 streaming contribution!!

JessicaMak and marcobrunodev streaming contribution!!

alfredotigolo, mandrakenk and Okbatgames chatting contribution!!

jitspoe, Endesga and Fearitself streaming contribution!!

jmcmorris (Jason Morris, SiegeGames) streaming contribution!!

tomrandall streaming Ludum contribution!!

DonHo streaming contribution!!

vue_equalizer (fpga_guru) chatting contribution!!

Thiagovgamg chatting contribution!!

Naysayer88 and friends contribution!!

lelandkwong streaming contribution!!

Goldbargames streaming contribution!!

Bytakos (bytakos) streaming contribution!!

Endesga streaming contribution!!

strager streaming contribution!!

jitspoe streaming contribution!!

Ash_F0x and JessicaMak streaming contribution!!

WTSRetro/SpiffyDane and Myrkee streaming contribution!!

Ninja and friends streaming contribution!!

erald_guri chatting contribution!!

lastmiles streaming farwest contribution!!

rw_grim streaming contribution!!

AdamCYounis streaming contribution!!

unicorndev (and Friends: metrix and Zultralord) streaming and chatting contribution!!

Dunno (P4ndaExpress) chatting and streaming contribution!!

Zorchenhimer streaming contribution!!

lasteveq4 C++ interest chat contriubtion!!

cecilphillip and clarkio @"Microsoft Developer" streaming contribution!!

oijtx streaming contribution!!

diegobrando_linux (Bl4ck_gookoo) chatting contribution!!

jhovgaard streaming contribution!!

Klay4_ chatting contribution!!

HonestDanGames streaming contribution!!

NorthSeaHero streaming contribution!!

Trainwreckstv and friends streaming contribution!!

togglebit, GexYT and GoPirateSoftware streaming contribution!!

taiyoinoue, RetroMMO, OfficialAndyPyro and david_joffe streaming contribution!!

Tjienta streaming contribution!!

ThePrimeagen streaming contribution!!

ghostfasm (epryka, liefasm, Vladimir - Vova from Russia) streaming contribuion!!

Jaxstyle and friends streaming contribution!!

EduardRFS streaming contribution!!

qpHalcy0n (and Friends: DragonSlayer0531) streaming and Linux and C++ development contribution!!

Melchizedek6809 and btcfly streaming contribution!!

Llama0x0 and sov_l chatting contribution!!

LLamabit chatting contribution!!

heroseh gamedev streaming contribution!!

TaleLearnCode streaming contribution!!

hvalen_hvalborg112 streaming contribution!!

Caro and lus contribution!!

Kebabskal streaming contribution!!

klintcsgo (CS GO: Counter-Strike Global Offensive) streaming contribution!!

løtchoh friendship contribution!!

KASPERPURE (Super Mario 64) streaming contribution!!

SomewhatAccurate C++ streaming contribution!!

TurtleWolfe chatting, hosting and raid contribution!!

ultristors chatting contribution!!

Listening to Bryan Adams, Westlife, Shayne Ward, MLTR, Backstreet Boys, Boyzone - Best Love Songs Ever by Relax Song at YouTube!!

-- hi5 contribution...!!

at Windows 10 Pro at eddie machine remotely from bilbo machine running Windows 10 Pro!!

+++ aura/user/notify_icon.cpp
@@ -78,7 +78,7 @@ namespace user
    {
 
       //return m_plistener->_get_notification_area_action_count();
-      return m_notifyiconitema.get_count();
+      return (int) m_notifyiconitema.get_count();
 
    }
